+++ configobject/objecttypes/notificationcommand/notificationcommand.go
@@ -32,7 +32,7 @@ type NotificationCommand struct {
 	NameCi             *string `json:"name_ci"`
 	ZoneId             string  `json:"zone_id"`
 	Command            string  `json:"command"`
-	Timeout            float32 `json:"timeout"`
+	Timeout            float64 `json:"timeout"`
 }
 
 func NewNotificationCommand() connection.Row {
